add refactoring test

+++ src/main/java/com/example/credit_service_project/generator/AccountGenerator.java
@@ -27,7 +27,7 @@ public class AccountGenerator {
         return LocalDate.now();
     }
 
-    public LocalDate LocalDateCreateClosingDate(Long yearsAmount) {
+    public LocalDate LocalDateCreateClosingDate(Integer yearsAmount) {
         return createOpeningDay().plusYears(yearsAmount);
     }
